Split paging params building into smaller methods for easy overriding.
